+++ Changelog.md
@@ -11,10 +11,12 @@
 - Ensure use of Intimidation skill is detected as an Approved Action for Intimidate.
 - #416: Allow **Compendium Search** of Items to filter based on cosm, traits, secondaryAxiom, category (incl. Implant Type).
 - (A bug in core Foundry prevents searching within Item fields in the Item sidebar: https://github.com/foundryvtt/foundryvtt/issues/1330 - a one-line patch in foundry.js could get it working)
+- Display highest size and vulnerable modifier in the Skill Test dialog.
 
 ### Bug Fixes
 
 - An error related to `approvedActions` would sometimes occur when rolling skill checks outside of combat.
+- Attacks against Vehicles should work once again.
 
 ### Translations